moved bitset to frontier folder
