--- cpp/oneapi/dal/backend/primitives/bitset.hpp
@@ -1,1 +1,0 @@
-#include "oneapi/dal/backend/primitives/bitset/bitset.hpp"
