+++ cpp/oneapi/dal/backend/primitives/frontier/frontier.hpp
@@ -4,7 +4,7 @@
 #include "oneapi/dal/backend/common.hpp"
 #include "oneapi/dal/graph/detail/common.hpp"
 #include "oneapi/dal/graph/detail/container.hpp"
-#include "oneapi/dal/backend/primitives/bitset.hpp"
+#include "oneapi/dal/backend/primitives/frontier/bitset.hpp"
 #include "oneapi/dal/backend/primitives/ndarray.hpp"
 
 namespace oneapi::dal::backend::primitives {
